Rename new_lobby to lobby
