+++ src/pages/lobby.tsx
@@ -11,9 +11,9 @@ import { useFunctions } from "../hooks/useFunctions";
  * COMPONENTS
  */
 import Button from "../components/buttons/button";
-import RoomList from "../components/new_lobby/RoomList";
+import RoomList from "../components/lobby/RoomList";
 
-import LobbyPlayerCard from "../components/new_lobby/lobby-player-card";
+import LobbyPlayerCard from "../components/lobby/lobby-player-card";
 
 const Lobby = () => {
   const { playerAccount } = useAuth();
